// hub/frontend/lib/api/generated/model/eventRuleRead.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.24
 */
import type { EventRuleReadActionConfig } from './eventRuleReadActionConfig';
import type { EventRuleReadConditions } from './eventRuleReadConditions';
import type { EventRuleReadFailurePolicy } from './eventRuleReadFailurePolicy';

export interface EventRuleRead {
  actionConfig: EventRuleReadActionConfig;
  actionType: string;
  conditions: EventRuleReadConditions;
  createdAt: string;
  createdByUserId: string | null;
  description: string;
  eventTypes: string[];
  failurePolicy: EventRuleReadFailurePolicy;
  id: string;
  isEnabled: boolean;
  lastTriggeredAt: string | null;
  name: string;
  ownerOrganizationId: string | null;
  ownerUserId: string | null;
  signingSecret?: string | null;
  updatedAt: string;
}